requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djoser==2.2.3
drf-spectacular==0.27.1
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
